# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "My Brand-Backend API Gabriel's Portifolio",
      "version": "1.0.0",
      "description": "API documentation for My Brand application(Gabriel's Portifolios)"
    },
    
    "servers": [
      {
        "url": "//localhost:3000",
        "description": "Gabriel's Brand server"
      }
    ],
    "components": {
        "securitySchemes": {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      },
    
    "paths": {
      "/api/user/signup": {
        "post": {
          "summary": "Sign up a new user",
          "tags": ["User-Authentication"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" },
                    "confirmPassword": { "type": "string" }
                  },
                  "required": ["firstName", "lastName", "email", "password","confirmPassword"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User signed up successfully"
            },
            "400": {
              "description": "Bad request - invalid input data"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
       "/api/user/signin": {
        "post": {
          "summary": "Sign in a user",
          "tags": ["User-Authentication"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User signed in successfully"
            }
          }
        }
      },
      
      "/api/user/{userId}": {
        "delete": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "summary": "Delete a user by ID",
          "tags": ["User-Authentication"],
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "description": "ID of the user to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User deleted successfully"
            }
          }
        },
        "put": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "summary": "Update a user by ID",
          "tags": ["User-Authentication"],
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "description": "ID of the user to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User updated successfully"
            }
          }}
        },
        "/api/user": {
            "get": {
              "security": [
                {
                  "BearerAuth": []
                }
              ],
              "summary": "Get all users",
              "tags": ["User-Authentication"],
              "responses": {
                "200": {
                  "description": "Users retrieved successfully"
                }
              }
            }
        
      },
     
      "/api/message": {
        "post": {
          "summary": "Create a new message",
          "tags": ["Contact-Message"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string" },
                    "email": { "type": "string" },
                    "message": { "type": "string" }
                    
                  },
                  "required": ["username", "email", "message"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Message created successfully"
            }
          }
        },
        "get": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "summary": "Get all messages",
          "tags": ["Contact-Message"],
          "responses": {
            "200": {
              "description": "Messages retrieved successfully"
            }
          }
        }
      },
      "/api/message/{messageId}": {
        
        "put": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "summary": "Update a message by ID",
          "tags": ["Contact-Message"],
          "parameters": [
            {
              "name": "messageId",
              "in": "path",
              "required": true,
              "description": "ID of the message to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string" },
                    "email": { "type": "string" },
                    "message": { "type": "string" }
                    
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Message updated successfully"
            }
          }
        },
        "delete": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "summary": "Delete a message by ID",
          "tags": ["Contact-Message"],
          "parameters": [
            {
              "name": "messageId",
              "in": "path",
              "required": true,
              "description": "ID of the message to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Message deleted successfully"
            }
          }
        }
      },
      "/api/blog": {
        "post": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "tags": ["Blogs"],
          "summary": "Create a new blog",
          "operationId": "createBlog",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Title of the blog",
                      "example": "Sample Title"
                    },
                    "description": {
                      "type": "string",
                      "description": "Description of the blog",
                      "example": "This is a sample description"
                    },
                    "image": {
                      "type": "string",
                      "format": "binary",
                      "description": "Image file of the blog"
                    }
                  },
                  "required": ["title", "description", "image"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Blog created successfully"
            }
          }
        },
      
      
        "get": {
          "tags": ["Blogs"],
          "summary": "Get all blogs",
          "operationId": "getAllBlogs",
          "responses": {
            "200": {
              "description": "List of blogs"
            }
          }
        }
      },
      
      
      
     
      "/api/blog/{id}": {
        "put": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "tags": ["Blogs"],
          "summary": "Update an existing blog",
          "operationId": "updateBlog",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "New title of the blog"
                    },
                    "description": {
                      "type": "string",
                      "description": "New description of the blog"
                    },
                    "image": {
                      "type": "string",
                      "description": "New image URL of the blog"
                    }
                  }
                }
              }
            }
          },
          
          "responses": {
            "200": {
              "description": "Updated blog"
            },
            "404": {
              "description": "Blog not found"
            }
          }
        
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["Blogs"],
        "summary": "Delete a blog",
        "operationId": "deleteBlog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the blog to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted successfully"
          },
          "404": {
            "description": "Blog not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
        "get": {
          "tags": ["Blogs"],
          "summary": "Get a blog by ID",
          "operationId": "getBlogById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to retrieve",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Blog found"
            }
          }
        }
      },
      "/api/blog/{id}/like": {
        "post": {
          "tags": ["Blog-Likes & Dislikes"],
          "summary": "Like a blog",
          "operationId": "likeBlog",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to like",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Blog liked successfully"
            }
          }
        }
      },
      "/api/blog/{id}/dislike": {
        "post": {
          "tags": ["Blog-Likes & Dislikes"],
          "summary": "Dislike a blog",
          "operationId": "dislikeBlog",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to dislike",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Blog disliked successfully"
            }
          }
        }
      },
      "/api/blog/{id}/comments": {
        "post": {
          "tags": ["Blog-Comments"],
          "summary": "Create a new comment on a blog",
          "operationId": "createComment",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to comment on",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": {
                      "type": "string",
                      "description": "Full name of the commenter"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email of the commenter"
                    },
                    "comment": {
                      "type": "string",
                      "description": "Comment text"
                    }
                  },
                  "required": ["fullName", "email", "comment"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Comment added successfully"
            },
            "404": {
              "description": "Blog not found"
            }
          }
        
      },
      
        "get": {
          "tags": ["Blog-Comments"],
          "summary": "Get all comments for a blog",
          "operationId": "getCommentsForBlog",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the blog to get comments for",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of comments"
            }
          }
        }
      },
      "/api/comments": {
        "get": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "tags": ["Blog-Comments"],
          "summary": "Get all comments",
          "operationId": "getAllComments",
          "responses": {
            "200": {
              "description": "List of comments"
            }
          }
        }
      },
      
      "/api/blog/{blogId}/comments/{commentId}": {
        "delete": {
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "tags": ["Blog-Comments"],
          "summary": "Delete a comment from a blog",
          "operationId": "deleteComment",
          "parameters": [
            {
              "name": "blogId",
              "in": "path",
              "required": true,
              "description": "ID of the blog from which to delete the comment",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "commentId",
              "in": "path",
              "required": true,
              "description": "ID of the comment to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Comment deleted successfully"
            }
          }
        }
      }
    }}
  
  